refactor(vault): drop dead DepositFlowStep re-export, btcWalletOverride, and DEPOSIT_SUCCESS_MESSAGE
- Remove the 'Re-export for convenience' of DepositFlowStep from
  depositStepHelpers; redirect SigningProgress, DepositProgressView,
  and ResumeDepositContent to import directly from
  @/hooks/deposit/depositFlowSteps. Update the test mock accordingly.
- Drop btcWalletOverride from usePayoutSigningState and the
  phaseTrackingWallet wrapper from ResumeSignContent — the resume
  payout flow now just renders SIGN_PAYOUTS while signing and
  ARTIFACT_DOWNLOAD with isWaiting=true on completion.
- Remove DEPOSIT_SUCCESS_MESSAGE; the new-deposit flow parks on
  ARTIFACT_DOWNLOAD and never reaches COMPLETED, so the success
  banner never rendered.
